test(memory): consolidate page-table and paging test harnesses
