+++ memory/page-table-generic/tests/copy.rs
@@ -10,28 +10,6 @@ const ROOT_ENTRY_SIZE: usize = 1 << 39;
 const KERNEL_SPACE_BASE: usize = 0xffff_8000_0000_0000;
 const KERNEL_IMAGE_BASE: usize = 0xffff_ffff_8000_0000;
 
-#[derive(Clone, Copy)]
-struct CanonicalT4kL4;
-
-impl TableMeta for CanonicalT4kL4 {
-    type P = PteImpl;
-
-    const PAGE_SIZE: usize = PAGE_SIZE;
-    const LEVEL_BITS: &'static [usize] = &[9, 9, 9, 9];
-    const MAX_BLOCK_LEVEL: usize = 3;
-
-    fn canonicalize_vaddr(vaddr: VirtAddr) -> VirtAddr {
-        let address = vaddr.as_usize() & ((1usize << 48) - 1);
-        VirtAddr::from_usize(if address & (1usize << 47) == 0 {
-            address
-        } else {
-            address | !((1usize << 48) - 1)
-        })
-    }
-
-    fn flush(_vaddr: Option<VirtAddr>) {}
-}
-
 fn map_page<A: FrameAllocator>(page_table: &mut PageTable<T4kL4, A>, vaddr: usize, paddr: usize) {
     page_table
         .map_page(
@@ -132,31 +110,6 @@ fn cloned_missing_root_entries_preserve_boot_only_kernel_mappings() {
         user.translate_phys(VirtAddr::from_usize(KERNEL_IMAGE_PAGE)),
         Ok(PhysAddr::from_usize(0x40_0000))
     );
-}
-
-#[test]
-fn walker_reports_canonical_high_addresses() {
-    const KERNEL_PAGE: usize = 0xffff_8000_0020_0000;
-
-    let mut page_table = PageTable::<CanonicalT4kL4, Fram4k>::new(Fram4k).unwrap();
-    page_table
-        .map_page(
-            VirtAddr::from_usize(KERNEL_PAGE),
-            PhysAddr::from_usize(0x40_0000),
-            0x1000,
-            MappingFlags::READ.into(),
-        )
-        .unwrap();
-
-    let mappings = page_table
-        .walk(
-            VirtAddr::from_usize(0xffff_8000_0000_0000),
-            VirtAddr::from_usize(usize::MAX),
-        )
-        .filter(|entry| entry.is_final_mapping)
-        .collect::<Vec<_>>();
-    assert_eq!(mappings.len(), 1);
-    assert_eq!(mappings[0].vaddr, VirtAddr::from_usize(KERNEL_PAGE));
 }
 
 #[test]
